split sidebar to components
