+++ src/lib/components/ChatMode.svelte
@@ -7,64 +7,26 @@
   import { g } from "$lib/global.svelte";
 
   import { slide } from "svelte/transition";
-  import { getProfile } from "$lib/profile.svelte";
   import { derivePromise, navigate, resolveLeafId } from "$lib/utils.svelte";
   import { Category, Channel } from "@roomy-chat/sdk";
-  import toast from "svelte-french-toast";
-  import { user } from "$lib/user.svelte";
   import { outerWidth } from "svelte/reactivity/window";
 
   let isMobile = $derived((outerWidth.current || 0) < 640);
 
   let sidebarAccordionValues = $state(["channels", "threads"]);
 
-  // Navigate to first channel in space if we do not have a channel selected.
-  $effect(() => {
-    if (!page.params.channel && !page.params.thread) {
-      (async () => {
-        if (!g.space) return;
 
-        for (const item of await g.space.sidebarItems.items()) {
-          const category = item.tryCast(Category);
-          const channel = item.tryCast(Channel);
-          if (category) {
-            for (const channel of await category.channels.items()) {
-              return navigate({
-                space: page.params.space!,
-                channel: channel.id,
-              });
-            }
-          } else if (channel) {
-            return navigate({
-              space: page.params.space!,
-              channel: channel.id,
-            });
-          }
-        }
-      })();
-    }
+  let availableThreads = derivePromise([], async () => {
+    const active = g.channel;
+    if (!active || !active.matches(Channel)) return [];
+    return ((await active.threads?.items()) || []).filter(
+      (x) => !x.softDeleted,
+    );
   });
 
-  let availableThreads = derivePromise([], async () =>
-    ((await g.space?.threads.items()) || []).filter((x) => !x.softDeleted),
-  );
+  let { categories, channels } = $props();
 
-  let categories = derivePromise([], async () => {
-    if (!g.space) return [];
-    return (await g.space.sidebarItems.items())
-      .map((x) => x.tryCast(Category) as Category)
-      .filter((x) => !!x);
-  });
 
-  let sidebarItems = derivePromise([], async () => {
-    if (!g.space) return [];
-    return await g.space.sidebarItems.items();
-  });
-
-  let channels = derivePromise([], async () =>{
-    if (!g.space) return [];
-    return await g.space.channels.items();
-  })
 
   let showNewCategoryDialog = $state(false);
   let newCategoryName = $state("");
@@ -79,87 +41,6 @@
     g.space.commit();
 
     showNewCategoryDialog = false;
-  }
-
-  let saveSpaceLoading = $state(false);
-  let showSpaceSettings = $state(false);
-  let newSpaceHandle = $state("");
-  let spaceNameInput = $state("");
-  let bannedHandlesInput = $state("");
-  let verificationFailed = $state(false);
-  $effect(() => {
-    if (!g.space) return;
-    if (!showSpaceSettings) {
-      spaceNameInput = g.space.name;
-      newSpaceHandle = g.space?.handles.get(0) || "";
-      verificationFailed = false;
-      saveSpaceLoading = false;
-      Promise.all(
-        Object.keys(g.space.bans.toJSON()).map((x) => getProfile(x)),
-      ).then(
-        (profiles) =>
-          (bannedHandlesInput = profiles.map((x) => x.handle).join(", ")),
-      );
-    }
-  });
-  async function saveBannedHandles() {
-    if (!g.space || !user.agent) return;
-    const bannedIds = (
-      await Promise.all(
-        bannedHandlesInput
-          .split(",")
-          .map((x) => x.trim())
-          .filter((x) => !!x)
-          .map((x) => user.agent!.resolveHandle({ handle: x })),
-      )
-    ).map((x) => x.data.did);
-    g.space.bans.clear();
-    for (const ban of bannedIds) {
-      g.space.bans.set(ban, true);
-    }
-    g.space.commit();
-    showSpaceSettings = false;
-  }
-  async function saveSpaceName() {
-    if (!g.space) return;
-    g.space.name = spaceNameInput;
-    g.space.commit();
-  }
-  async function saveSpaceHandle() {
-    if (!g.space) return;
-    saveSpaceLoading = true;
-
-    if (!newSpaceHandle) {
-      g.space.handles.clear();
-      g.space.commit();
-      saveSpaceLoading = false;
-      showSpaceSettings = false;
-      toast.success("Saved space with without handle.", {
-        position: "bottom-right",
-      });
-      return;
-    }
-
-    try {
-      const id = await resolveLeafId(newSpaceHandle);
-      if (!id) {
-        verificationFailed = true;
-        saveSpaceLoading = false;
-        return;
-      }
-      g.space.handles.clear();
-      g.space.handles.push(newSpaceHandle);
-      g.space.commit();
-      saveSpaceLoading = false;
-      showSpaceSettings = false;
-      toast.success("Space handle successfully verified & saved!", {
-        position: "bottom-right",
-      });
-    } catch (e) {
-      saveSpaceLoading = false;
-      verificationFailed = true;
-      console.error(e);
-    }
   }
 
   let showNewChannelDialog = $state(false);
@@ -203,125 +84,11 @@
 
 <nav
   class={[
-    !isMobile &&
-      "min-h-0 p-0 overflow-y-auto",
+    !isMobile && "min-h-0 p-0 overflow-y-auto",
     "p-0 flex flex-col gap-4 w-full",
   ]}
   style="scrollbar-width: thin;"
 >
-  <div class="flex justify-between">
-    {#if g.isAdmin}
-      <Dialog title="Space Settings" bind:isDialogOpen={showSpaceSettings}>
-        {#snippet dialogTrigger()}
-          <Button.Root
-            title="Space Settings"
-            class="btn w-full justify-start join-item text-base-content"
-          >
-            <Icon icon="lucide:settings" class="size-6" />
-          </Button.Root>
-        {/snippet}
-
-        <form onsubmit={saveSpaceName} class="flex flex-col gap-3">
-          <label class="input w-full">
-            <span class="label">Name</span>
-            <input bind:value={spaceNameInput} placeholder="My Space" />
-          </label>
-          <Button.Root class="btn btn-primary w-full">Save Name</Button.Root>
-        </form>
-        <form class="flex flex-col gap-6" onsubmit={saveSpaceHandle}>
-          <h2 class="font-bold text-xl">Handle</h2>
-          <div class="flex flex-col gap-2">
-            <p>
-              Space handles are created with DNS records and allow your space to
-              be reached at a URL like <code
-                >https://roomy.chat/-/example.org</code
-              >.
-            </p>
-            {#if !!newSpaceHandle}
-              {@const subdomain = newSpaceHandle
-                .split(".")
-                .slice(0, -2)
-                .join(".")}
-              <p>
-                Add the following DNS record to your DNS provider to use the
-                domain as your handle.
-              </p>
-              <div class="max-w-full overflow-x-auto min-w-0">
-                <table class="table text-[0.85em]">
-                  <thead>
-                    <tr>
-                      <th>Type</th>
-                      <th>Host</th>
-                      <th>Value</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <tr>
-                      <td>TXT</td>
-                      <td>
-                        _leaf{subdomain ? "." + subdomain : ""}
-                      </td>
-                      <td>
-                        "id={g.space.id}"
-                      </td>
-                    </tr>
-                  </tbody>
-                </table>
-              </div>
-            {:else}
-              <p>Provide a domain to see which DNS record to add for it.</p>
-            {/if}
-          </div>
-          <label class="input w-full">
-            <span class="label">Handle</span>
-            <input bind:value={newSpaceHandle} placeholder="example.org" />
-          </label>
-
-          {#if verificationFailed}
-            <div role="alert" class="alert alert-error">
-              <span
-                >Verification failed. It may take several minutes before DNS
-                records are propagated. If you have configured them correctly
-                try again in a few minutes.</span
-              >
-            </div>
-          {/if}
-
-          <Button.Root class="btn btn-primary" bind:disabled={saveSpaceLoading}>
-            {#if saveSpaceLoading}
-              <span class="loading loading-spinner"></span>
-            {/if}
-            {!!newSpaceHandle ? "Verify" : "Save Without Handle"}
-          </Button.Root>
-        </form>
-
-        <form class="flex flex-col gap-4" onsubmit={saveBannedHandles}>
-          <h2 class="font-bold text-xl">Bans</h2>
-
-          <div>
-            <input class="input w-full" bind:value={bannedHandlesInput} />
-            <div class="flex flex-col">
-              <span class="mx-2 mt-1 text-sm"
-                >Input a list of handles separated by commas.</span
-              >
-              <span class="mx-2 mt-1 text-sm"
-                >Note: the ban is "best effort" right now. The Roomy alpha is
-                generally insecure.</span
-              >
-            </div>
-          </div>
-
-          <Button.Root
-            class="btn btn-primary w-full"
-            bind:disabled={saveSpaceLoading}
-          >
-            Save Bans
-          </Button.Root>
-        </form>
-      </Dialog>
-    {/if}
-  </div>
-
   {#if g.isAdmin}
     <menu class="menu p-0 w-full justify-between join join-vertical">
       <Dialog title="Create Channel" bind:isDialogOpen={showNewChannelDialog}>
@@ -407,29 +174,6 @@
           {/snippet}
         </Accordion.Content>
       </Accordion.Item>
-      {#if availableThreads.value.length > 0}
-        <div class="divider my-0"></div>
-        <Accordion.Item value="threads">
-          <Accordion.Header>
-            <Accordion.Trigger
-              class="cursor-pointer flex w-full items-center justify-between mb-2 uppercase text-xs font-medium text-base-content"
-            >
-              <h3>Threads</h3>
-              <Icon
-                icon="basil:caret-up-solid"
-                class={`size-4 transition-transform duration-150 ${sidebarAccordionValues.includes("threads") && "rotate-180"}`}
-              />
-            </Accordion.Trigger>
-          </Accordion.Header>
-          <Accordion.Content>
-            {#snippet child({ open }: { open: boolean })}
-              {#if open}
-                {@render threadsSidebar()}
-              {/if}
-            {/snippet}
-          </Accordion.Content>
-        </Accordion.Item>
-      {/if}
     </Accordion.Root>
   </ToggleGroup.Root>
 </nav>
@@ -441,8 +185,9 @@
 {#snippet channelsSidebar()}
   <div transition:slide class="flex flex-col gap-4">
     <!-- Category and Channels -->
-    {#each sidebarItems.value.filter((x) => !x.softDeleted) as item}
+    {#each channels.value as item}
       {@const category = item.tryCast(Category)}
+      <!-- TODO: ToggleGroup's by category -->
       {#if category}
         <Accordion.Root type="single" value={item.name}>
           <Accordion.Item value={item.name}>
@@ -552,13 +297,18 @@
           </h3>
         </ToggleGroup.Item>
       {/if}
+      {#if item.id === page.params.channel}
+        <div class="opacity-80 pl-2 -mt-4">
+          {@render threadsSidebar(4)}
+        </div>
+      {/if}
     {/each}
   </div>
 {/snippet}
 
-{#snippet threadsSidebar()}
-  <div transition:slide class="flex flex-col gap-4">
-    {#each availableThreads.value as thread}
+{#snippet threadsSidebar(limit = Infinity)}
+  <div class="flex flex-col gap-1">
+    {#each availableThreads.value.filter((_, i) => i < limit) as thread}
       <ToggleGroup.Item
         onclick={() =>
           navigate({ space: page.params.space!, thread: thread.id })}
